cman build: fix different bits
adapt top level Makefile to cope with the libccs/libcman build loop.

cman_tool needs to know where to find libccs now in order to link properly.

bump libcman soname to 4.0 since it's not ABI compliant with the old one.

fix libcman build to include missing -I and -L

keep default of cmanincdir to point to libcman.h for now or nothing builds.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>