cman: Add some (hopefully) helpful comments
One of the functions I hope that this libcman will provide is
a guide to migrating from libcman to corosync interfaces. So, while
this is a functional libcman it also serves as a sort of document
as to where the libcman calls have gone (into cfg, votequorum etc)
and how to call them.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>