rgmanager: Fix nofailback when service is in 'starting' state
If a failover domain has "nofailback" enabled a service that is
"starting" state would still failback to the original.  If a
service takes a while to come up and is in the "starting" state
and the original node comes back online while it is in that state
the service will be failed back.

Signed-off-by: Shane Bradley <sbradley@redhat.com>
Signed-off-by: Lon Hohberger <lhh@redhat.com>