fenced: don't ignore victim_done messages for reduced victims
When a victim is "reduced" (i.e. fenced skips fencing it because it
rejoins the cluster cleanly before fenced fences it), it is immediately
removed from the list of victims, before the "victim_done" message is
sent for it.  The victim_done message updates the time of the last
successful fencing operation for a failed node.

The code that processes received victim_done messages was ignoring the
message for the reduced victim because the node couldn't be found in
the victims list.  This caused the latest fencing information to not be
recorded for the node, causing dlm_controld to wait indefinately for
fencing to complete for the reduced victim.

The fix is to simply record the information from a victim_done message
even if the node is not in the victims list.

bz 678704

Acked-by: Ryan O'Hara <rohara@redhat.com>
Signed-off-by: David Teigland <teigland@redhat.com>