rgmanager: Pause during exit if we stopped services
A difference between rgmanager 1.9.x and later versions is that
they rely on openais/corosync for messaging.  This messaging is
quite reliable and has proved useful.

However, one drawback is that if you very quickly stop rgmanager
and corosync/cman, the other nodes in the cluster can not restart
services because message traffic is interrupted for the duration
of the token timeout.

There is no simple solution to this problem.  Rgmanager could
(in theory) find new placements for services prior to stopping,
but this is a large amount of design work; it was never designed
to run policies in the exit path.

A far simpler idea is to simply give the other nodes time to
restart services.

NOTE: This solution does not and can not work with central
      processing mode.

Resolves: rhbz#619468

Signed-off-by: Lon Hohberger <lhh@redhat.com>