rgmanager: Fix segfault on config reload
This issue was initially triggered by rpm package which wrongly
set executable flag on resource-script.metadata file in
/usr/share/cluster/. After running "cman_tool version -r" cluster
crashed, last log entry was:

  rgmanager[47960]: Loading Service Data

Further investigation has shown, that this behaviour is triggered
always if any runnable file in /usr/share/cluster provides output
which is not well formed XML when run with the meta-data operation.

This behaviour is probably caused by the wrong usage of
xmlInitParser()/XmlCleanUp() parser.

Resolves: rhbz#692895

Signed-off-by: Yevheniy Demchenko <zheka@uvt.cz>
Reviewed-by: Lon Hohberger <lhh@redhat.com>
Tested-by: Lon Hohberger <lhh@redhat.com>