cman: allow late close of stderr file descriptor and free resources
In case debug is enabled, and cman is executed via init script, we leak
a fd associated to a file that will grow indefinetely in size.

We address this situation by adding a very specific cman_tool startup
option associated to init script, to tell cman to perform a late
close of stderr and freeing correctly all resources.

Resolves: rhbz#740385

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>