fenced: fix handling of startup partition merge
The victims created on each side of a partition are cleared after
a merge in the receive_complete function, which is meant to clear
"initial victims".  Clearing the victims is the correct end result,
but the code arrives there through an unintended shortcut.  Change
the code so it clears the victims in a "more correct", and probably
safer way:

2 is blocked doing startup fencing
1 joins fence domain
partition between 1,2
1 sees confchg for partition, adds victim 2 (and sets init_victim
  per this patch, since 1 hasn't finished a start cycle yet)
partition removed
2 completes startup fencing
2 sees confchg for partition, adds victim 1
1 sees confchg for merge, adds node 2
2 sees confchg for merge, adds node 1
1 processing the merge confchg
2 reduces victim 1 from partition, since 1 had no state (had not yet
  completed a start cycle)
2 processing the merge confchg
1,2 finish start cycle for merge confchg
2 sends complete for merge confchg
1 clears victim 2 in receive_complete because it set init_victim

bz 750314

Signed-off-by: David Teigland <teigland@redhat.com>