dlm_controld: fix handling of startup partition merge
Nack messages are not matched correctly with previous change
structs.  Normal start messages should not be matched with
changes that preceeded the node's join time, but nack messages
should, so the match_change() function needs to distinguish.

This can occur when two nodes are joining a lockspace, but
neither has completed when a cluster partition+merge happens.

bz 750314

Signed-off-by: David Teigland <teigland@redhat.com>