rgmanager: OCF_NOT_INSTALLED is OK if we stop when stopped
Scripts often live in file systems which are mounted by
rgmanager.  This patch makes rgmanager ignore the
OCF_NOT_INSTALLED error returned from the script agent if
we attempt to stop when a start failed.

Furthermore, don't mark the resource as failed if a start
failure occurs due to the program not existing.

The combination of these create the ability for a missing
script to move to a node where the script exists, or be
placed in to the stopped state.

Resolves: rhbz#768146

Signed-off-by: Lon Hohberger <lhh@redhat.com>
Reviewed-by: Adam Drew <adrew@redhat.com>
Tested-by: Adam Drew <adrew@redhat.com>