cman: Improve quorum timer handling how quorum timers work
Previously, the timer was simply once per quorumd poll interval.

This led to a possibility of qdiskd (or other programs using the
CMAN quorum API) to continue as 'alive' until a time period greater
than the poll interval.

With this patch, the quorum device timer is reset each time the
program using the quorum API.

Resolves: rhbz#759603

Signed-off-by: John Ruemker <jruemker@redhat.com>
Reviewed-by: Lon Hohberger <lhh@redhat.com>
Reviewed-by: Christine Caulfield <ccaulfie@redhat.com>