rgmanager: Fix for services stuck in recovery
Patch from John Ruemker <jruemker@redhat.com>:

"When starting rgmanager throughout the cluster around the same
time, multiple nodes may end up acting as the "root" for a particular
service. If that service happens to fail on startup, you can end up
with each of those nodes sending remote-start requests around the
cluster.  Eventually the service will get stuck in a recovering state,
and cannot be modified in any way with clusvcadm. The only remedy we've
found is to kill rgmanager and start it back up.

Acked-by: Lon Hohberger <lhh@redhat.com>
Signed-off-by: Ryan McCabe <rmccabe@redhat.com>