Use new corosync confdb API to remove string limit.
The original corosync confdb API had a 256 byte limit on all
value strings returned. A new API was added that did not have this
restriction, and this commit cases (mainly) libccs and others to
use it so those benefits can be passed down to other cluster
services.

This means that cluster will now only compile against
corosync libraries that have the *_typed2 API calls.

rhbz#847234

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>