fenced: fix potential tight loop reading /dev/zero
The FD returned from cman is not meant to be stored anywhere - this is documented in the header
file (which is the official source of documentation for the API). Under some circumstances cman
can return the FD for /dev/zero (which is always active) and if the client application stores i
this instead of the one it expects then it will loop forever.

It's probably VERY hard to reproduce this, there would need to be some data coming from cman
quite soon after startup for it to happen, though it has been seen.

See rhbz#856214

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>