cman|fenced: Fix node killing in case of a 2node cluster that suffers brief network outage
This patch fixes a rare but nasty condition in cman and fenced. In a 2node cluster
if the network splits for a period of time longer than the token timeout but
shorter than the time needed to fence a node then both nodes can send 'kill'
packets to the other with the effect that both nodes' cmans will quit
leaving no operational cluster.

This patch adds a check for a 2node cluster and only sends a 'kill' packet
to the node with the higher nodeid thus ensuring a predictable response
to such events and ensuring that services can continue to run.

The cman part of this patch is only active if Disallowed is enabled,
which is rare; the fenced part is more likely to be invoked

Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>
Acked-By: Fabio M. Di Nitto <fdinitto@redhat.com>