xml: ccs_update_schema: be verbose about extraction fail
Previously, the distillation of resource-agents' metadata could fail
from unexpected reasons without any evidence ever being made, unlike
in case of fence-agents.  Also "no metadata" and "issue with their
extraction" will allegedly yield the same outcome, so it is reflected
in the comments being emitted to the schema for both sorts of agents.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>