liblogthread: Avoid potetntial race when changing log files
There is a potential for a crash when a log file is changed or
re-opened as the fclose can happen in the main thread while the
writing thread is about to output data.

This patch avoids that by allocating a new fp for the new file and
only closing the old fp* later one when the thread holds the lock.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>